Hotfix: move common load methods into load/util.
`src/load/webpack` is no longer including the values imported from
`src/load` in webpack builds. Separating the common methods between the
two into a shared module circumvents this issue.
