+++ tests/unit/load.ts
@@ -4,7 +4,8 @@ import * as assert from 'intern/chai!assert';
 import * as registerSuite from 'intern!object';
 import * as sinon from 'sinon';
 import has from '../../src/has';
-import load, { isPlugin, useDefault } from '../../src/load';
+import load from '../../src/load';
+import { useDefault } from '../../src/load/util';
 import mockPlugin from '../support/load/plugin-default';
 import global from '../../src/global';
 
@@ -15,46 +16,6 @@ const suite: any = {
 
 	before() {
 		return load('tests/support/load/a', 'tests/support/load/b', 'tests/support/load/c');
-	},
-
-	isPlugin() {
-		assert.isFalse(isPlugin(null));
-		assert.isFalse(isPlugin(2));
-		assert.isFalse(isPlugin([]));
-		assert.isFalse(isPlugin(/\s/));
-		assert.isFalse(isPlugin({}));
-		assert.isTrue(isPlugin({
-			load() {}
-		}));
-	},
-
-	useDefault: {
-		'single es6 module'() {
-			assert.strictEqual(useDefault({
-				'__esModule': true,
-				'default': 42
-			}), 42, 'The default export should be returned.');
-		},
-
-		'single non-es6 module'() {
-			const module = { value: 42 };
-			assert.deepEqual(useDefault(module), module, 'The module itself should be returned.');
-		},
-
-		'all es6 modules'() {
-			const modules = [ 42, 43 ].map((value: number) => {
-				return { '__esModule': true, 'default': value };
-			});
-			assert.sameMembers(useDefault(modules), [ 42, 43 ], 'The default export should be returned for all modules.');
-		},
-
-		'mixed module types'() {
-			const modules: any[] = [ 42, 43 ].map((value: number) => {
-				return { '__esModule': true, 'default': value };
-			});
-			modules.push({ value: 44 });
-			assert.sameDeepMembers(useDefault(modules), [ 42, 43, { value: 44 } ]);
-		}
 	},
 
 	'global load'(this: any) {
@@ -198,16 +159,6 @@ if (has('host-node')) {
 			result.then(def.callback(function ([ a, b ]: [ any, any ]) {
 				assert.deepEqual(a, { 'default': 'A', one: 1, two: 2 });
 				assert.deepEqual(b, { 'default': 'B', three: 3, four: 4 });
-			}));
-		},
-
-		'useDefault resolves es modules'(this: any) {
-			const def = this.async(5000);
-
-			const result: Promise<any[]> = nodeRequire(path.join(buildDir, 'tests', 'support', 'load', 'node')).succeedDefault;
-			result.then(def.callback(function ([ a, b ]: [ any, any ]) {
-				assert.deepEqual(a, 'A');
-				assert.deepEqual(b, 'B');
 			}));
 		},
